% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Calculates the curvature at each point of a series}
\usage{
curvature(points)
}
\arguments{
\item{points}{a sf object with multiple features with POINT geometry}
}
\value{
a sf object with multiple features with POINT geometry and attributes corresponding to curvilinear coordinate (S) and curvature (C)
}
\description{
Calculates the curvature at each point of a series
}
\examples{
points=st_read(system.file("data/points.shp", package="morphRiver"))
rasterDEM=raster(system.file("data/rasterDEM.tif", package="morphRiver"))
plot(rasterDEM)
plot(points,
     add=TRUE, col=1)
rcurv=curvature(points) \%>\% points_to_linestring()
plot(rcurv["C"],
    add=TRUE, lwd=3)

# Curvature as longitudinal signal
library(ggplot2)
ggplot(data=rcurv, aes(x=S,y=C))+
  geom_path()
}
