% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{elevation}
\alias{elevation}
\title{Calculates the elevation at each point of a series based on a DEM raster}
\usage{
elevation(points, rasterDEM)
}
\arguments{
\item{points}{a sf object with multiple features with POINT geometry}

\item{rasterDEM}{a raster object giving elevations}
}
\value{
a sf object with multiple features with POINT geometry and attributes corresponding to curvilinear coordinate (S) and elevation (Z))
}
\description{
Calculates the elevation at each point of a series based on a DEM raster
}
\examples{
points=st_read(system.file("data/points.shp", package="morphRiver"))
rasterDEM=raster(system.file("data/rasterDEM.tif", package="morphRiver"))
plot(rasterDEM)
plot(points,
     add=TRUE, col=1)
relev=elevation(points,rasterDEM) \%>\% points_to_linestring()
plot(relev["Z"],
     add=TRUE, lwd=3)
# Elevation as longitudinal signal
library(ggplot2)
ggplot(data=relev, aes(x=S,y=Z))+
  geom_path()
}
