% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_linestring.R
\name{points_to_linestring}
\alias{points_to_linestring}
\title{Calculates the curvature at each point of a series}
\usage{
points_to_linestring(points)
}
\arguments{
\item{points}{a sf object with multiple features with POINT geometry}
}
\value{
a sf object with multiple features with POINT geometry and attributes corresponding to curvilinear coordinate (S) and curvature (C)
}
\description{
Calculates the curvature at each point of a series
}
\examples{
plot(rasterDEM)
plot(points,
    add=TRUE, col=1)
plot(points_to_linestring(points),
    add=TRUE, col=2)
}
