% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{transects}
\alias{transects}
\title{Calculates the perpendicular transects to a series of points}
\usage{
transects(points, radius)
}
\arguments{
\item{points}{a sf object with multiple features with POINT geometry}

\item{radius}{the radius to consider for the transect}
}
\value{
a sf object with multiple features with LINESTRING geometry, corresponding to transects
}
\description{
Calculates the perpendicular transects to a series of points
}
\examples{
points=st_read(system.file("data/points.shp", package="morphRiver"))
rasterDEM=raster(system.file("data/rasterDEM.tif", package="morphRiver"))

plot(rasterDEM)
plot(points,
     add=TRUE, col=1)
plot(points_to_linestring(points),
     add=TRUE, col=2)
plot(transects(points,500),
     add=TRUE, col=4)
}
